test: RecentlyViewedArtworks
